% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire data set.}
\format{
A data set with 15 rows and two variables:
\describe{
\item{Distance}{Distrance from fire station in miles }
\item{Damage}{X1000 dollars in damages}
}
}
\usage{
fire
}
\description{
Fire data set.
}
\keyword{datasets}
